src/pyproject.toml
[tool.poetry]
name = "twitfix"
version = "0.1.0"
description = ""
authors = ["stormydragon", "robinuniverse", "et. al"]
include = [
    "static/*",
    "templates/*",
]

[tool.poetry.scripts]
twitfix = "twitfix.wsgi:main"

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.12.17"
twitter = "^1.19.3"
sanic = "^22.3.2"
Jinja2 = "^3.1.2"
Sanic-Cors = "^2.0.1"
requests = "^2.27.1"

# Local server deployment optionals
pymongo = { version = "^4.0.2", optional=true }
uWSGI = { version = "^2.0.20", optional=true }

# GCP Deployment optionals
google-cloud-firestore = { version = "^2.4.0", optional=true }
google-cloud-storage = { version = "^2.2.1", optional=true }

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
deploy-here = ["pymongo", "uWSGI"]
deploy-compose = ["pymongo"]
deploy-gcp = ["google-cloud-firestore", "google-cloud-storage"]
